simplify DOM element handling and testing
