+++ test/node/tests.js
@@ -63,42 +63,8 @@ const react = [
   }
 ];
 
-let domElements = [];
-if (typeof document !== 'undefined') {
-  const element1 = document.createElement('div');
-  const element2 = document.createElement('div');
-  const element3 = document.createElement('input');
-
-  domElements = [{
-    description: 'DOM elements',
-    tests: [
-      {
-        description: 'equal DOM elements',
-        value1: element1,
-        value2: element1,
-        equal: true
-      },
-      {
-        description: 'comparison of different elements',
-        value1: element1,
-        value2: element2,
-        equal: false
-      },
-      {
-        description: 'comparison of elements with different types',
-        value1: element1,
-        value2: element3,
-        equal: false
-      },
-    ]
-  }];
-}
-
-
-
-
 module.exports = {
   generic,
   react,
-  all: [...generic, ...react, ...domElements],
+  all: [...generic, ...react],
 };
